Move `cluster-panel` component to `shared` directory
