+++ modules/web/src/app/cluster/details/shared/cluster-panel/style.scss
