+++ modules/web/src/app/cluster/details/shared/cluster-panel/theme.scss
