+++ modules/web/src/assets/css/theme/_component.scss
@@ -14,7 +14,7 @@
 
 @import 'src/app/cluster/details/cluster/cni-version/theme';
 @import 'src/app/cluster/details/cluster/constraints/theme';
-@import 'src/app/cluster/details/cluster/machine-deployment-details/cluster-panel/theme';
+@import 'src/app/cluster/details/shared/cluster-panel/theme';
 @import 'src/app/cluster/details/shared/version-picker/theme';
 @import 'src/app/core/components/footer/theme';
 @import 'src/app/core/components/navigation/project/theme';
